+++ TDEngine2/plugins/VulkanGraphicsContext/include/CVulkanGraphicsContext.h
@@ -382,6 +382,8 @@ namespace TDEngine2
 			TDE2_API TPtr<IWindowSystem> GetWindowSystem() const override;
 
 			TDE2_API std::vector<U8> GetBackBufferData() const override;
+
+			TDE2_API VkDevice GetDevice() const;
 		protected:
 			DECLARE_INTERFACE_IMPL_PROTECTED_MEMBERS(CVulkanGraphicsContext)
 			TDE2_API CVulkanGraphicsContext(TPtr<IWindowSurfaceFactory> pWindowSurfaceFactory);